backend/requirements.txt
azure-functions
azure-identity
azure-mgmt-costmanagement
azure-mgmt-subscription
sqlalchemy
python-dateutil
